Potential fix for code scanning alert no. 39: Multiplication result converted to larger type

+++ include/linux/backing-dev.h
@@ -94,7 +94,7 @@ extern void wb_writeout_inc(struct bdi_writeback *wb);
 static inline unsigned long wb_stat_error(void)
 {
 #ifdef CONFIG_SMP
-	return nr_cpu_ids * WB_STAT_BATCH;
+	return (unsigned long)nr_cpu_ids * WB_STAT_BATCH;
 #else
 	return 1;
 #endif
